fix for KernelSU
KernelSU overlayfs has a small difference to Magisk that breaks this app.

Overlayfs does not preserve symlinks, so when the module mounts over /system/product (which is a symlink to /product), the symlink gets broken, and the result is an empty /system/product with only the debloated apps in it. The original /product is untouched, hence this app never worked for KernelSU.

This commit simply fixes that.